added main files
the actual code is found under Templates/index.html, the rest enables the webapp to work. If you are wondering why there is a python server and yet the app is written in JS, it's mostly because I never intended for this to be the final version and everything ended up working ok in JS. I might migrate it to Python some day, as it would allow better cross-browser compatibility.